+++ lib/src/rendering/sliver_sticky_header.dart
@@ -1,10 +1,10 @@
 import 'dart:math' as math;
 
 import 'package:flutter/foundation.dart';
+import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter_sticky_header/flutter_sticky_header.dart';
-import 'package:flutter_sticky_header/src/rendering/sticky_header_constraints.dart';
-import 'package:flutter_sticky_header/src/rendering/sticky_header_layout_builder.dart';
+import 'package:value_layout_builder/value_layout_builder.dart';
 
 /// A sliver with a [RenderBox] as header and a [RenderSliver] as child.
 ///
@@ -21,17 +21,28 @@ class RenderSliverStickyHeader extends RenderSliver with RenderSliverHelpers {
         _overlapsContent = overlapsContent,
         _reverse = reverse {
     this.header = header;
+    RenderObject? header,
+    RenderSliver? child,
+    bool overlapsContent: false,
+    bool sticky: true,
+    bool reverse: false,
+    StickyHeaderController? controller,
+  })  : _overlapsContent = overlapsContent,
+        _sticky = sticky,
+        _reverse = reverse,
+        _controller = controller {
+    this.header = header as RenderBox?;
     this.child = child;
   }
 
-  SliverStickyHeaderState _oldState;
-  double _headerExtent;
-  bool _isPinned;
+  SliverStickyHeaderState? _oldState;
+  double? _headerExtent;
+  late bool _isPinned;
 
   bool get overlapsContent => _overlapsContent;
   bool _overlapsContent;
+
   set overlapsContent(bool value) {
-    assert(value != null);
     if (_overlapsContent == value) return;
     _overlapsContent = value;
     markNeedsLayout();
@@ -46,82 +57,103 @@ class RenderSliverStickyHeader extends RenderSliver with RenderSliverHelpers {
     markNeedsLayout();
   }
 
+  bool get sticky => _sticky;
+  bool _sticky;
+
+  set sticky(bool value) {
+    if (_sticky == value) return;
+    _sticky = value;
+    markNeedsLayout();
+  }
+
+  StickyHeaderController? get controller => _controller;
+  StickyHeaderController? _controller;
+
+  set controller(StickyHeaderController? value) {
+    if (_controller == value) return;
+    if (_controller != null && value != null) {
+      // We copy the state of the old controller.
+      value.stickyHeaderScrollOffset = _controller!.stickyHeaderScrollOffset;
+    }
+    _controller = value;
+  }
+
   /// The render object's header
-  RenderBox get header => _header;
-  RenderBox _header;
-  set header(RenderBox value) {
-    if (_header != null) dropChild(_header);
+  RenderBox? get header => _header;
+  RenderBox? _header;
+
+  set header(RenderBox? value) {
+    if (_header != null) dropChild(_header!);
     _header = value;
-    if (_header != null) adoptChild(_header);
+    if (_header != null) adoptChild(_header!);
   }
 
   /// The render object's unique child
-  RenderSliver get child => _child;
-  RenderSliver _child;
-  set child(RenderSliver value) {
-    if (_child != null) dropChild(_child);
+  RenderSliver? get child => _child;
+  RenderSliver? _child;
+
+  set child(RenderSliver? value) {
+    if (_child != null) dropChild(_child!);
     _child = value;
-    if (_child != null) adoptChild(_child);
+    if (_child != null) adoptChild(_child!);
   }
 
   @override
   void setupParentData(RenderObject child) {
     if (child.parentData is! SliverPhysicalParentData)
-      child.parentData = new SliverPhysicalParentData();
+      child.parentData = SliverPhysicalParentData();
   }
 
   @override
   void attach(PipelineOwner owner) {
     super.attach(owner);
-    if (_header != null) _header.attach(owner);
-    if (_child != null) _child.attach(owner);
+    if (_header != null) _header!.attach(owner);
+    if (_child != null) _child!.attach(owner);
   }
 
   @override
   void detach() {
     super.detach();
-    if (_header != null) _header.detach();
-    if (_child != null) _child.detach();
+    if (_header != null) _header!.detach();
+    if (_child != null) _child!.detach();
   }
 
   @override
   void redepthChildren() {
-    if (_header != null) redepthChild(_header);
-    if (_child != null) redepthChild(_child);
+    if (_header != null) redepthChild(_header!);
+    if (_child != null) redepthChild(_child!);
   }
 
   @override
   void visitChildren(RenderObjectVisitor visitor) {
-    if (_header != null) visitor(_header);
-    if (_child != null) visitor(_child);
+    if (_header != null) visitor(_header!);
+    if (_child != null) visitor(_child!);
   }
 
   @override
   List<DiagnosticsNode> debugDescribeChildren() {
     List<DiagnosticsNode> result = <DiagnosticsNode>[];
     if (header != null) {
-      result.add(header.toDiagnosticsNode(name: 'header'));
+      result.add(header!.toDiagnosticsNode(name: 'header'));
     }
     if (child != null) {
-      result.add(child.toDiagnosticsNode(name: 'child'));
+      result.add(child!.toDiagnosticsNode(name: 'child'));
     }
     return result;
   }
 
   double computeHeaderExtent() {
     if (header == null) return 0.0;
-    assert(header.hasSize);
-    assert(constraints.axis != null);
+    assert(header!.hasSize);
     switch (constraints.axis) {
       case Axis.vertical:
-        return header.size.height;
+        return header!.size.height;
       case Axis.horizontal:
-        return header.size.width;
+        return header!.size.width;
     }
-    return null;
   }
 
-  double get headerLogicalExtent => overlapsContent ? 0.0 : _headerExtent;
+  double? get headerLogicalExtent => overlapsContent ? 0.0 : _headerExtent;
 
   @override
   void performLayout() {
@@ -137,10 +169,10 @@ class RenderSliverStickyHeader extends RenderSliver with RenderSliverHelpers {
     if (reverse) axisDirection = flipAxisDirection(axisDirection);
 
     if (header != null) {
-      header.layout(
-        new StickyHeaderConstraints(
-          state: _oldState ?? new SliverStickyHeaderState(0.0, false),
-          boxConstraints: constraints.asBoxConstraints(),
+      header!.layout(
+        BoxValueConstraints<SliverStickyHeaderState>(
+          value: _oldState ?? SliverStickyHeaderState(0.0, false),
+          constraints: constraints.asBoxConstraints(),
         ),
         parentUsesSize: true,
       );
@@ -148,12 +180,14 @@ class RenderSliverStickyHeader extends RenderSliver with RenderSliverHelpers {
     }
 
     // Compute the header extent only one time.
-    double headerExtent = headerLogicalExtent;
-    double headerPaintExtent = calculatePaintOffset(constraints, from: 0.0, to: headerExtent);
-    double headerCacheExtent = calculateCacheOffset(constraints, from: 0.0, to: headerExtent);
+    double headerExtent = headerLogicalExtent!;
+    final double headerPaintExtent =
+        calculatePaintOffset(constraints, from: 0.0, to: headerExtent);
+    final double headerCacheExtent =
+        calculateCacheOffset(constraints, from: 0.0, to: headerExtent);
 
     if (child == null) {
-      geometry = new SliverGeometry(
+      geometry = SliverGeometry(
           scrollExtent: headerExtent,
           maxPaintExtent: headerExtent,
           paintExtent: headerPaintExtent,
@@ -162,7 +196,7 @@ class RenderSliverStickyHeader extends RenderSliver with RenderSliverHelpers {
           hasVisualOverflow:
               headerExtent > constraints.remainingPaintExtent || constraints.scrollOffset > 0.0);
     } else {
-      child.layout(
+      child!.layout(
         constraints.copyWith(
           scrollOffset: math.max(0.0, constraints.scrollOffset - headerExtent),
           cacheOrigin: math.min(0.0, constraints.cacheOrigin + headerExtent),
@@ -172,20 +206,22 @@ class RenderSliverStickyHeader extends RenderSliver with RenderSliverHelpers {
         ),
         parentUsesSize: true,
       );
-      final SliverGeometry childLayoutGeometry = child.geometry;
+      final SliverGeometry childLayoutGeometry = child!.geometry!;
       if (childLayoutGeometry.scrollOffsetCorrection != null) {
-        geometry = new SliverGeometry(
+        geometry = SliverGeometry(
           scrollOffsetCorrection: childLayoutGeometry.scrollOffsetCorrection,
         );
         return;
       }
 
       final double paintExtent = math.min(
-          headerPaintExtent +
-              math.max(childLayoutGeometry.paintExtent, childLayoutGeometry.layoutExtent),
-          constraints.remainingPaintExtent);
+        headerPaintExtent +
+            math.max(childLayoutGeometry.paintExtent,
+                childLayoutGeometry.layoutExtent),
+        constraints.remainingPaintExtent,
+      );
 
-      geometry = new SliverGeometry(
+      geometry = SliverGeometry(
         scrollExtent: headerExtent + childLayoutGeometry.scrollExtent,
         paintExtent: paintExtent,
         layoutExtent: math.min(headerPaintExtent + childLayoutGeometry.layoutExtent, paintExtent),
@@ -197,48 +233,66 @@ class RenderSliverStickyHeader extends RenderSliver with RenderSliverHelpers {
         hasVisualOverflow: childLayoutGeometry.hasVisualOverflow,
       );
 
-      final SliverPhysicalParentData childParentData = child.parentData;
-      assert(constraints.axisDirection != null);
-      assert(constraints.growthDirection != null);
 
+      final SliverPhysicalParentData? childParentData =
+          child!.parentData as SliverPhysicalParentData?;
       final double childStartOffset = reverse ? headerExtent : headerPaintExtent;
-
       switch (axisDirection) {
         case AxisDirection.up:
-          childParentData.paintOffset = Offset.zero;
+          childParentData!.paintOffset = Offset.zero;
           break;
         case AxisDirection.right:
-          childParentData.paintOffset = Offset(childStartOffset, 0.0);
+          childParentData!.paintOffset = Offset(
+              calculatePaintOffset(constraints, from: 0.0, to: childStartOffset),
+              0.0);
           break;
         case AxisDirection.down:
-          childParentData.paintOffset = new Offset(0.0, childStartOffset);
+          childParentData!.paintOffset = Offset(0.0,
+              calculatePaintOffset(constraints, from: 0.0, to: childStartOffset1));
           break;
         case AxisDirection.left:
-          childParentData.paintOffset = Offset.zero;
+          childParentData!.paintOffset = Offset.zero;
           break;
       }
     }
 
     if (header != null) {
-      final SliverPhysicalParentData headerParentData = header.parentData;
-      final childScrollExtent = child?.geometry?.scrollExtent ?? 0.0;
-      double headerPosition = math.min(constraints.overlap,
-          childScrollExtent - constraints.scrollOffset - (overlapsContent ? _headerExtent : 0.0));
+      final SliverPhysicalParentData? headerParentData =
+          header!.parentData as SliverPhysicalParentData?;
+      final double childScrollExtent = child?.geometry?.scrollExtent ?? 0.0;
+      final double headerPosition = sticky
+          ? math.min(
+              constraints.overlap,
+              childScrollExtent -
+                  constraints.scrollOffset -
+                  (overlapsContent ? _headerExtent! : 0.0))
+          : -constraints.scrollOffset;
 
-      _isPinned = (constraints.scrollOffset + constraints.overlap) > 0.0 ||
-          constraints.remainingPaintExtent == constraints.viewportMainAxisExtent;
-      // second layout if scroll percentage changed and header is a RenderStickyHeaderLayoutBuilder.
-      if (header is RenderStickyHeaderLayoutBuilder) {
-        double scrollPercentage =
-            ((headerPosition - constraints.overlap).abs() / _headerExtent).clamp(0.0, 1.0);
+      _isPinned = sticky &&
+          ((constraints.scrollOffset + constraints.overlap) > 0.0 ||
+              constraints.remainingPaintExtent ==
+                  constraints.viewportMainAxisExtent);
 
-        SliverStickyHeaderState state = new SliverStickyHeaderState(scrollPercentage, _isPinned);
+      final double headerScrollRatio =
+          ((headerPosition - constraints.overlap).abs() / _headerExtent!);
+      if (_isPinned && headerScrollRatio <= 1) {
+        controller?.stickyHeaderScrollOffset =
+            constraints.precedingScrollExtent;
+      }
+      // second layout if scroll percentage changed and header is a
+      // RenderStickyHeaderLayoutBuilder.
+      if (header is RenderConstrainedLayoutBuilder<
+          BoxValueConstraints<SliverStickyHeaderState>, RenderBox>) {
+        double headerScrollRatioClamped = headerScrollRatio.clamp(0.0, 1.0);
+
+        SliverStickyHeaderState state =
+            SliverStickyHeaderState(headerScrollRatioClamped, _isPinned);
         if (_oldState != state) {
           _oldState = state;
-          header.layout(
-            new StickyHeaderConstraints(
-              state: _oldState,
-              boxConstraints: constraints.asBoxConstraints(),
+          header!.layout(
+            BoxValueConstraints<SliverStickyHeaderState>(
+              value: _oldState!,
+              constraints: constraints.asBoxConstraints(),
             ),
             parentUsesSize: true,
           );
@@ -247,18 +301,18 @@ class RenderSliverStickyHeader extends RenderSliver with RenderSliverHelpers {
 
       switch (axisDirection) {
         case AxisDirection.up:
-          headerParentData.paintOffset =
-              new Offset(0.0, geometry.paintExtent - headerPosition - _headerExtent);
+          headerParentData!.paintOffset = Offset(
+              0.0, geometry!.paintExtent - headerPosition - _headerExtent!);
           break;
         case AxisDirection.down:
-          headerParentData.paintOffset = new Offset(0.0, headerPosition);
+          headerParentData!.paintOffset = Offset(0.0, headerPosition);
           break;
         case AxisDirection.left:
-          headerParentData.paintOffset =
-              new Offset(geometry.paintExtent - headerPosition - _headerExtent, 0.0);
+          headerParentData!.paintOffset = Offset(
+              geometry!.paintExtent - headerPosition - _headerExtent!, 0.0);
           break;
         case AxisDirection.right:
-          headerParentData.paintOffset = new Offset(headerPosition, 0.0);
+          headerParentData!.paintOffset = Offset(headerPosition, 0.0);
           break;
       }
     }
@@ -266,20 +320,25 @@ class RenderSliverStickyHeader extends RenderSliver with RenderSliverHelpers {
 
   @override
   bool hitTestChildren(SliverHitTestResult result,
-      {@required double mainAxisPosition, @required double crossAxisPosition}) {
-    assert(geometry.hitTestExtent > 0.0);
-    if (header != null && mainAxisPosition - constraints.overlap <= _headerExtent) {
-      return hitTestBoxChild(BoxHitTestResult.wrap(SliverHitTestResult.wrap(result)), header,
-              mainAxisPosition: mainAxisPosition - constraints.overlap,
-              crossAxisPosition: crossAxisPosition) ||
+      {required double mainAxisPosition, required double crossAxisPosition}) {
+    assert(geometry!.hitTestExtent > 0.0);
+    if (header != null &&
+        mainAxisPosition - constraints.overlap <= _headerExtent!) {
+      return hitTestBoxChild(
+            BoxHitTestResult.wrap(SliverHitTestResult.wrap(result)),
+            header!,
+            mainAxisPosition: mainAxisPosition - constraints.overlap,
+            crossAxisPosition: crossAxisPosition,
+          ) ||
           (_overlapsContent &&
               child != null &&
-              child.geometry.hitTestExtent > 0.0 &&
-              child.hitTest(result,
-                  mainAxisPosition: mainAxisPosition - childMainAxisPosition(child),
+              child!.geometry!.hitTestExtent > 0.0 &&
+              child!.hitTest(result,
+                  mainAxisPosition:
+                      mainAxisPosition - childMainAxisPosition(child),
                   crossAxisPosition: crossAxisPosition));
-    } else if (child != null && child.geometry.hitTestExtent > 0.0) {
-      return child.hitTest(result,
+    } else if (child != null && child!.geometry!.hitTestExtent > 0.0) {
+      return child!.hitTest(result,
           mainAxisPosition: mainAxisPosition - childMainAxisPosition(child),
           crossAxisPosition: crossAxisPosition);
     }
@@ -287,15 +346,19 @@ class RenderSliverStickyHeader extends RenderSliver with RenderSliverHelpers {
   }
 
   @override
-  double childMainAxisPosition(RenderObject child) {
-    if (child == header) return _isPinned ? 0.0 : -(constraints.scrollOffset + constraints.overlap);
+  double childMainAxisPosition(RenderObject? child) {
+    if (child == header)
+      return _isPinned
+          ? 0.0
+          : -(constraints.scrollOffset + constraints.overlap);
     if (child == this.child)
-      return calculatePaintOffset(constraints, from: 0.0, to: headerLogicalExtent);
-    return null;
+      return calculatePaintOffset(constraints,
+          from: 0.0, to: headerLogicalExtent!);
+    return 0;
   }
 
   @override
-  double childScrollOffset(RenderObject child) {
+  double? childScrollOffset(RenderObject child) {
     assert(child.parent == this);
     if (child == this.child) {
       return _headerExtent;
@@ -306,23 +369,25 @@ class RenderSliverStickyHeader extends RenderSliver with RenderSliverHelpers {
 
   @override
   void applyPaintTransform(RenderObject child, Matrix4 transform) {
-    assert(child != null);
-    final SliverPhysicalParentData childParentData = child.parentData;
+    final SliverPhysicalParentData childParentData =
+        child.parentData as SliverPhysicalParentData;
     childParentData.applyPaintTransform(transform);
   }
 
   @override
   void paint(PaintingContext context, Offset offset) {
-    if (geometry.visible) {
-      if (child != null && child.geometry.visible) {
-        final SliverPhysicalParentData childParentData = child.parentData;
-        context.paintChild(child, offset + childParentData.paintOffset);
+    if (geometry!.visible) {
+      if (child != null && child!.geometry!.visible) {
+        final SliverPhysicalParentData childParentData =
+            child!.parentData as SliverPhysicalParentData;
+        context.paintChild(child!, offset + childParentData.paintOffset);
       }
 
-      // The header must be draw over the sliver.
+      // The header must be drawn over the sliver.
       if (header != null) {
-        final SliverPhysicalParentData headerParentData = header.parentData;
-        context.paintChild(header, offset + headerParentData.paintOffset);
+        final SliverPhysicalParentData headerParentData =
+            header!.parentData as SliverPhysicalParentData;
+        context.paintChild(header!, offset + headerParentData.paintOffset);
       }
     }
   }